Insert from File to DB Complete
Was losing precision on data going into DB, changed script from
"decimal" to "money", precision was rounded to an acceptable range.
Able to insert into the database.
Updated logging to append to file.
Added stubs to allow table creation, and dropping <- should be done in
the DB using stored procs.